Crate pallet_evm_precompile_author_mapping

Precompile to interact with pallet author mapping through an evm precompile.

Structs

AuthorMappingPrecompile

A precompile to wrap the functionality from pallet author mapping.

GetKeysSize

Bound for the keys size. Pallet will check that the size exactly matches, but we want to bound the parser to not accept larger bytes.

Enums

AuthorMappingPrecompileCall